test(eval): 修复独立安装环境中的跨进程测试导入
